Added MUtilBase and allowed NULL in MDateTime objects
For the API to be a bit simpler, where all these objects are used, there is a
base object for all - MUtilBase - which supportd the "toString(), isNull() and
isValid() functions.

Also the MDateTime need to support the "null" concept if to be used in the
Macaroni suite, so included that.