Fixed bug where MDate("now") created unwanted time portion.
The new MDate("now") should just return today, and the
internal date property needs to be midnight of today (00:00 hours).
It contined the current time.  This could make adding
a day to it come out with the incorrect date as it adds a whole
bunch to get around daylight saving in this scenario..